+++ docs/conf.py
@@ -2,7 +2,7 @@ from javaproperties import __version__
 
 project   = 'javaproperties'
 author    = 'John T. Wodder II'
-copyright = '2016-2019 John T. Wodder II'
+copyright = '2016-2020 John T. Wodder II'
 
 extensions = [
     'sphinx.ext.autodoc',